module reorganization and refactoring
